backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
slowapi==0.1.9  # API速率限制

# 文档处理
pypdf==5.1.0
python-docx==1.1.2
markdown==3.7
pypandoc==1.14
python-magic==0.4.27

# 向量嵌入
sentence-transformers==2.7.0
flagembedding==1.2.7
openai==1.57.0
anthropic==0.42.0

# 向量数据库
faiss-cpu==1.8.0
pymilvus==2.4.9
qdrant-client==1.12.1
marshmallow==3.21.3  # 修复pymilvus的marshmallow版本冲突

# LLM
openai==1.57.0
anthropic==0.42.0

# 本地 LLM 支持
transformers>=4.35.0
torch>=2.0.0
huggingface_hub>=0.19.0
accelerate>=0.24.0

# 工具库
numpy>=1.24.0,<2.0.0
scipy==1.14.1
aiofiles==24.1.0
python-dotenv==1.0.1

# 文档解析增强
tqdm==4.66.4
transformers==4.46.1
html2text==2024.2.26
python-pptx==0.6.23
openpyxl==3.1.5
pandas==2.2.2

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5

# RAGFlow 风格切分功能
tiktoken>=0.5.0  # 精确的 Token 计算
beautifulsoup4>=4.12.0  # HTML 解析（用于 HTML 切分）
lxml>=4.9.0  # BeautifulSoup 的依赖